refactor(engine): move test util to upper level for reuse
